1. Fixed Sidebar navigation  2. Added Some inner-shadow to some components 3. Readjusted the date 4. Location marks earned some animations.

+++ client/src/components/ui/Header.tsx
@@ -56,7 +56,7 @@ export default function Header() {
     const today = new Date()
     today.setDate(today.getDate() - 1);
     const options: Intl.DateTimeFormatOptions = {
-      month: 'numeric',
+      month: 'long',
       day: 'numeric'
     }
     return `${today.toLocaleDateString("en-US", options)}`
@@ -66,7 +66,7 @@ export default function Header() {
     const today = new Date()
     today.setDate(today.getDate() + 1);
     const options: Intl.DateTimeFormatOptions = {
-      month: 'numeric',
+      month: 'long',
       day: 'numeric'
     }
     return `${today.toLocaleDateString("en-US", options)}`
